+++ pkg/physarum/particle.go
@@ -1,8 +1,8 @@
 package physarum
 
 type Particle struct {
-	X float64
-	Y float64
-	A float64
-	C int
+	X float32
+	Y float32
+	A float32
+	C uint32
 }
